svn merge -r 1966:1976 svn://svn.minisip.org/var/svn/minisip/branches/ppc_evc4_cesc/minisip .
libmutil compiles in MS EVC++ 4.0




git-svn-id: svn://svn.minisip.org/minisip/trunk@1977 63dae60f-c5d3-0310-92a3-fcbf00261c72

+++ libmutil/include/libmutil/itoa.h
@@ -34,11 +34,12 @@
 #ifndef ITOA_H
 #define ITOA_H
 
+#include <libmutil/libmutil_config.h>
+
 #include<string>
+#include<libmutil/mtypes.h>
 
-#include<libmutil_config.h>
-
-LIBMUTIL_API std::string itoa(int i);
+LIBMUTIL_API std::string itoa(int64_t i);
 
 #endif
